files formated

+++ src/interfaces/core/ICdai.sol
@@ -4,83 +4,63 @@ pragma solidity ^0.8.24;
 import {IERC20Metadata} from "@openzeppelin/contracts/interfaces/IERC20Metadata.sol";
 
 interface ICdai is IERC20Metadata {
-    function mint(uint mintAmount) external returns (uint);
+    function mint(uint256 mintAmount) external returns (uint256);
 
-    function redeem(uint redeemTokens) external returns (uint);
+    function redeem(uint256 redeemTokens) external returns (uint256);
 
-    function redeemUnderlying(uint redeemAmount) external returns (uint);
+    function redeemUnderlying(uint256 redeemAmount) external returns (uint256);
 
-    function borrow(uint borrowAmount) external returns (uint);
+    function borrow(uint256 borrowAmount) external returns (uint256);
 
-    function repayBorrow(uint repayAmount) external returns (uint);
+    function repayBorrow(uint256 repayAmount) external returns (uint256);
 
-    function repayBorrowBehalf(
-        address borrower,
-        uint repayAmount
-    ) external returns (uint);
+    function repayBorrowBehalf(address borrower, uint256 repayAmount) external returns (uint256);
 
-    function transfer(address dst, uint amount) external returns (bool);
+    function transfer(address dst, uint256 amount) external returns (bool);
 
-    function transferFrom(
-        address src,
-        address dst,
-        uint amount
-    ) external returns (bool);
+    function transferFrom(address src, address dst, uint256 amount) external returns (bool);
 
-    function approve(address spender, uint amount) external returns (bool);
+    function approve(address spender, uint256 amount) external returns (bool);
 
-    function allowance(
-        address owner,
-        address spender
-    ) external view returns (uint);
+    function allowance(address owner, address spender) external view returns (uint256);
 
-    function balanceOf(address owner) external view returns (uint);
+    function balanceOf(address owner) external view returns (uint256);
 
-    function balanceOfUnderlying(address owner) external returns (uint);
+    function balanceOfUnderlying(address owner) external returns (uint256);
 
-    function getAccountSnapshot(
-        address account
-    ) external view returns (uint, uint, uint, uint);
+    function getAccountSnapshot(address account) external view returns (uint256, uint256, uint256, uint256);
 
-    function borrowRatePerBlock() external view returns (uint);
+    function borrowRatePerBlock() external view returns (uint256);
 
-    function supplyRatePerBlock() external view returns (uint);
+    function supplyRatePerBlock() external view returns (uint256);
 
-    function totalBorrowsCurrent() external returns (uint);
+    function totalBorrowsCurrent() external returns (uint256);
 
-    function borrowBalanceCurrent(address account) external returns (uint);
+    function borrowBalanceCurrent(address account) external returns (uint256);
 
-    function borrowBalanceStored(address account) external view returns (uint);
+    function borrowBalanceStored(address account) external view returns (uint256);
 
-    function exchangeRateCurrent() external returns (uint);
+    function exchangeRateCurrent() external returns (uint256);
 
-    function exchangeRateStored() external view returns (uint);
+    function exchangeRateStored() external view returns (uint256);
 
-    function getCash() external view returns (uint);
+    function getCash() external view returns (uint256);
 
-    function accrueInterest() external returns (uint);
+    function accrueInterest() external returns (uint256);
 
-    function seize(
-        address liquidator,
-        address borrower,
-        uint seizeTokens
-    ) external returns (uint);
+    function seize(address liquidator, address borrower, uint256 seizeTokens) external returns (uint256);
 
     function underlying() external view returns (address);
 
     function totalReserves() external view returns (uint256);
 
-    function _addReserves(uint addAmount) external returns (uint);
+    function _addReserves(uint256 addAmount) external returns (uint256);
 
-    function _setPendingAdmin(
-        address payable newPendingAdmin
-    ) external returns (uint);
+    function _setPendingAdmin(address payable newPendingAdmin) external returns (uint256);
 
-    function _acceptAdmin() external returns (uint);
+    function _acceptAdmin() external returns (uint256);
 
-    function _setReserveFactor(
-        uint newReserveFactorMantissa
-    ) external returns (uint);
+    function _setReserveFactor(uint256 newReserveFactorMantissa) external returns (uint256);
 
-    function _reduceReserves(uint reduceAmount) external returns (uint);
+    function _reduceReserves(uint256 reduceAmount) external returns (uint256);
 }
